storage: improve CRC checking on read
When reading an object, no longer rely on the presence of a
Content-Length header to determine whether to check a CRC. That was
always fragile, and in fact it doesn't work under HTTP 1.

Instead, check two new signals to determine whether the bytes read
differ from the bytes stored in GCS (and thus will give a different checksum):

- http.Response.Uncompressed, set by the Go http stack when it uncompresses
  incoming gzip data.

- The value of the X-Google-Gfe-Response-Body-Transformations header,
  where the server reports what it did with the data.

Also, remove the checkedCRC field from the reader and the test that
checks it, since now we always check the CRC if we determine from the
response that we should.

Change-Id: I5306906b3505a3309fe4f1f3e8b3e401b05c056a
Reviewed-on: https://code-review.googlesource.com/29570
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>