httpreplay: fix 1.6, 1.7 build by providing a stub httpreplay package
We added build tags to httpreplay that let it only build on 1.8 and fowards. Unfortunately
this means that the compiler sees no files in the package in 1.6 and 1.7, which means
when we go test ./... we get "no buildable Go source files in x/y/z/httpreplay".

This CL fixes that by providing an empty stub.

Change-Id: Id855cd9c376709449e1fef62e730d85c7477f73a
Reviewed-on: https://code-review.googlesource.com/29770
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>