spanner: unflake TestSessionRecycle
Currently, TestSessionRecycle sets the pool config to MinOpened: 1, MaxIdle: 2.
The idea is that MaxIdle being 2 means it's _possible_ for 2 sessions to be
spun up. However, the test itself is taking and subsequently recycling
sessions, meaning that if everything is working correctly there should only
ever be 1 session, despite there being room for 2. The test asserts that after
20 take/recycle iterations, there is only 1 session.

However, occasionally the test fails because there are two sessions. The reason
this happens is that in the background, spanner's session pool maintainer is
also futzing with the number of open sessions. Specifically, if there are
_zero_ sessions it will try to increase up to MinOpened (1). If this occurs
at the same time as the initial take, which also creases a session if there
is none, then we end up with 2 sessions. That number never decreases because
MaxIdle=2, meaning 2 open sessions is acceptable.

This interaction is not really a data race: each takes the same lock, so the
interaction does happen sequentially. The problem is that the two have
different conditionals guarding the behavior. The maintainer looks at whether
it needs to be involved: for example, is NumOpened < MinOpened? The take call,
however, looks at a lot less things - it essentially only cares whether
NumOpened < MaxOpened. So if they both get to the point where they're about to
commit to creating a session, and maintainer goes first, 2 will be created
(take does not care if it goes above MinOpened). OTOH, if take goes first,
only 1 will be created (maintainer _does_ care about going above MinOpened).

I gave some thought to changing the behavior so that take is more strict about
increasing session count, but I choose the relax the test assertions instead
because these interactions seem reasonable and OK.

This CL allows 1 or 2 sessions at the end. To compensate for this - to make
sure we're still checking that recycling works - the MaxIdle has been increased
to 5. If recycling did not work, we would see 2<x<=5 sessions, and the test
would fail.

Change-Id: I9f92a77bb005620d6d6021d41338b041dca2ab2d
Reviewed-on: https://code-review.googlesource.com/c/36970
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Eno Compton <enocom@google.com>