firestore: switch to apiv1
This switches firestore from the apiv1beta1 API to the apiv1 API. It
also switches several RPCs to include "/documents" in parent paths:

In the v1beta1 API, firestore accidentally allowed documents to be
queried with a path that did not include /documents. In v1, this was
disallowed. Where this matters:

- Docrefs: but these already have /documents added as part of their
initialization step.
- Collrefs: most collrefs are the child of a docref, which already
adds /documents as described above. However, top-level collrefs don't
already have this. So, in this CL we add /documents to top-level collref
parents.
- Places where we use client.Path() directly. In these places we add
/documents after client.Path(), keeping client.Path as-is to avoid
confusion.

Also in this CL we pull in https://github.com/googleapis/google-cloud-common/pull/271.

Change-Id: Ibdf4de9ee59def2fe9b60ece9c0b1e06b00a0e3c
Reviewed-on: https://code-review.googlesource.com/c/37413
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>