bigquery: add support for the GEOGRAPHY type.
BigQuery supports a GEOGRAPHY datatype, which is functionally
a string in a Well Known Text (WKT) format.  This change adds
support for interacting with the GEOGRAPHY type.

For consuming data from a query/table, the NullGeography type
is provided (as it is for other supported datatypes) for disambiguating
the default value from a missing (null) value.

Type inference from native go types is not supported, as there's not
a clear WKT type available.  We may add a string type alias in the
future if there's demand.

This change does NOT add GEOGRAPHY type support for query.
Added doc comments in params.go indicating the way to handle this.

Change-Id: I2b25fcdfd56fb9d3a774d4c93d13812ad19e6c72
Reviewed-on: https://code-review.googlesource.com/c/37810
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>