datastore: adds sentinel value DetectProjectID
When creating a Datastore client via NewClient, the sentinel value
DetectProjectID can be used in place of a project ID. This instructs
NewClient to use the given credentials to find the project ID.

There are limitations to this approach. For example a JWT does not have
the project ID encoded. NewClient is unable to be too helpful in
detecting where the user went wrong as option.ClientOption is opaque and
does now allow for inspection. DetectProjectID has some docs for a user
here.

NewClient defers to the environment variable DATASTORE_PROJECT_ID if it
is set. Otherwise, it will parse and retrieve the credentials via
google.golang.org/api/transport.Creds(). This implies that if the user
is using default credentials, then NewClient will read the project ID
from them.

Change-Id: Id364e8586d7e39a3d8385ac5746417b084a49dcf
Reviewed-on: https://code-review.googlesource.com/c/38450
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Chris Broadfoot <cbro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>