pubsub: add Subscription.PushConfig.AuthenticationMethod
Adds Subscription.PushConfig.AuthenticationMethod which is
a one-of field hence the use of an interface
`AuthenticationMethod`. This field is optional
and only part of a closed alpha.
The only currently supported AuthenticationMethod is
"OIDCToken" (OpenID Connect Token), which has been added.

This change adds an example too as well roundtrip tests.

Updates #1361

Change-Id: I6de49ad57dbe813bec66b84671ffca6094b6ca5c
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/40771
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Alex Hong <hongalex@google.com>