bigquery: move RequirePartitionFilter outside of TimePartitioning.
When first added to BigQuery, RequirePartitionFilter was added as a
field of TimePartitioning.

As BigQuery has more partitioning features coming, the ability to
require partition filters has been moved up to the table level
instead of residing within the specific partitioning configuration.

I have marked TimePartitioning.RequirePartitionFilter deprecated.
You can still manipulate it, but the top level field is the preferred
mechanism.

You can access the property via TableMetadata.RequirePartitionFilter
and update it directly in TableMetadataToUpdate.  In the backend
it is effectively a cloned field.

The backend behavior is undefined for cases where you attempt to set
Table.RequirePartitionFilter and
Table.TimePartition.RequirePartitionFilter to different values
as part of the same operation (table create, table update, etc).

Change-Id: I35f4e050e4c4e25b6a7c97f6bbaae7447654e8f0
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/40830
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>