pubsub: make ExpirationPolicy's type optional.Duration
This CL changes:
* Subscription.ExpirationPolicy
* SubscriptionToChange.ExpirationPolicy
from
   time.Duration
to
   optional.Duration

which is a somewhat backwards compatible change for this package,
but if users are comparing .ExpirationPolicy, they'll now have
to use optional.ToDuration but also to never expire an
ExpirationPolicy to Ttl, users will have to explicitly pass
in a typed time.Duration(0) value.

In comparison, the GCP UI and gcloud cli have a sentinel value "never".

Updates #1361

Change-Id: I4f992fcc4f2c81c9b593336524c9b0e71df916b5
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/41211
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>