internal: only check incompatible changes on presubmit
There doesn't seem to be a good reason to do so for continuous builds, yet
it's somehow failing our build.

Also takes the test timeout from 60m to 30m now that we've taken care of
problematic tests. The average time seems to be about 20m, so this should
be enough.

Change-Id: Ie1e76ca4a16c8c882f6420255baa469154430dba
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/42311
Reviewed-by: Emmanuel Odeke <emmanuel@orijtech.com>