logging: carve out module
This CL adds the second module to cloud.google.com/go, making it a
multi-module repository. This directory was chosen as a guinea pig due
to the fact that it sees fairly low amount of changes, is a client that has
value to carve out:

- It allows us to communicate client stability with semver, instead of all
  clients being contain in a mega v0.Y.Z module.
- It allows us to separate the per-client dependencies.

This CL will be tagged v0.43.0 and logging/v1.0.0.

Change-Id: I0fb297284aa544bd01098b34f035280edaf7907e
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/41530
Reviewed-by: Chris Broadfoot <cbro@google.com>